refactor(logger): switch to go-logr/zerologr
Use optimized implementation instead of json.Marshal with e.Interface()
